Uploading base protocol module and a few textures
Renderer not working yet - for some reason my window is black when there should be a cube there - If anyone knows why this is happening, respond to the issue for it. If you are interested in BlueMC, you may email me at bluemc.game@gmail.com to ask to work on the project.